apps for Linux

+++ spmv/src/spmvbench/smvm_timing_results.c
@@ -40,7 +40,7 @@ smvm_save_timing_results (struct SMVM_timing_results* p_results,
 			  const int nnzb, const int num_trials, 
 			  const double t_median, const double t_min, 
 			  const double t_max, const double mflops, 
-			  const int num_loads, const int num_stores)
+			  const long num_loads, const long num_stores)
 {
   if (p_results == NULL)
     {
